Restructure of Error
This commit modifies many of the functionalities of the library to
utilize the Error enumeration as a return. The library now has near-full support
for error returns. This branch has also tested all configurations of the
comparator, and is now a fit release candidate.

Master will be pulled up to date with this commit.